+++ RedPandaIDE/src/utils/parser.cpp
@@ -27,8 +27,6 @@ void resetCppParser(std::shared_ptr<CppParser> parser, int compilerSetIndex)
     // Configure parser
     parser->resetParser();
     parser->setEnabled(true);
-    parser->setParseGlobalHeaders(true);
-    parser->setParseLocalHeaders(true);
 
     // Set options depending on the current compiler set
     if (compilerSetIndex<0) {