+++ NEWS.md
@@ -15,6 +15,7 @@ Red Panda C++ Version 3.5
   - fix: Can't correctly complete folder name ending with "\\" in #include preprocessors.
   - change: Option "Auto clear symbol table when editor's hidden"  default to false.
   - fix: Block comments are not correctly replaced with space char.
+  - change: Remove "parse system headers" and "parse local headers" options.
 
 Red Panda C++ Version 3.4
   - enhancement: Auto hide option "Auto clear parsed symbols when editor hidden" if "editors share one parser" is unchecked.